there is no more localStorage.js services. Only cookies. Refactoring in that way.
