+++ vite-project/src/components/LoginButtons.jsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom'
 import { toggleConnected } from '../redux.js'
 import { useSelector } from 'react-redux'
-import { deleteLoginCookie } from '../services/localStorage.js'
+import { deleteLoginCookie } from '../services/cookies.js'
 
 export default function LoginButtons () {
     const userLogin = useSelector((state) => state.userLogin)
