+++ vite-project/src/redux.js
@@ -1,13 +1,11 @@
 import { configureStore, createSlice } from '@reduxjs/toolkit'
-import { getLoginCookie } from './services/localStorage.js'
+import { getLoginCookie } from './services/cookies.js'
 
 const cookies = getLoginCookie()
 
 let connectedInitialValue = false
 let tokenInitialValue = null
 let firstnameInitialValue = null
-
-console.log(cookies)
 
 if (cookies.token) {
     connectedInitialValue = true
